Refactor api call response dispatch logic
Previously, we had trouble figuring out which aws command we were receiving response for. The reason is because we used clojure spec to conform the data received from aws after an api call to determine how we should format the output on the chat channel.
The problem with this approach though is that some api call responses may "conform" to more than one spec (e.g: Group creation response, and Get Group Info response which contain both a field :Group). So one way to fix the problem is to make sure that the most specific specs are on the top of the dispatch stack (which is cond-based) but this is more of a hack.
A much better approach is to 'annotate' the response itself with metadata containing the aws call name being made, since the caller function knows which aws command it is receiving response for. The advantage of using metadata, compared to for e.g assoc-ing a field in the aws call response itself, is that we do not pollute the response shape by having an additional field hence we do not have to change our specs.